Move app auth handler to module
The app auth handler needs a lot of functionality, so putting it in its
own module will allow a cleaner division between app auth handler
helpers and other Splinter functionality.
