+++ daemon/src/splinter/app_auth_handler/error.rs
