+++ daemon/src/splinter/mod.rs
@@ -17,5 +17,4 @@
 
 pub mod app_auth_handler;
 pub mod batch_submitter;
-pub mod error;
 pub mod event;
